requirements.txt
streamlit==1.24.0
tensorflow==2.12.0
numpy==1.24.3
pillow==9.5.0
pandas==2.0.2
plotly==5.15.0
requests==2.31.0